Refactor codebase and added a new button.
1. On the function selectGridCell(), I remove
unnecessary foreach() for each event listeners.
Instead, I just put all my event listeners
on one single foreach() for optimization and
smooth performance.

I also change how I call the
getGridColors() by calling the function outside
the foreach() method.

This avoids calling all columns every single
time an event listener
was triggered and also getting the colors of each
cells using the getGridColors(), which can
cause lag on the browser.

2. I added a new button to change the current
behavior of selecting a background color of
the cell. Instead of "input" as the event
listener, I change it to "click" so that
when the user selects a color on the color picker,
they will need to click the "Fill" button
in order to apply the color. This is a simple way
to prevent lag on the browser since it only
triggers the color change once per
click instead of continuously while dragging
the picker.