These are the improvements I made for the Dark shading feature on the codebase.
1. function darkenCell() - This function is used
to darken the cell color.

2. function darkenRGB() - This function is used
to darken the RGB color(the rgb color from the
darkenCell function) every time a cell was
clicked or shaded using the Dark Shading button.

3. Additional changes - fixed the boolean logic
for event listeners of eraseCellButton and
darkShadingButton in order for the buttons
to not execute when the other one was clicked.