These are the changes I made for the css mobile UI design for the grid cells.
1. Refactor the function createGrid - I added a new variable: containerStyles, containerWidth and
containerHeight to set the width and height of the grid dynamically based on the number of cells and the
size of the cells, so that the grid is responsiveand dynamically adjust
their size based on the size of the screen (media queries).

2. I added a new event listener for the window resize event
to recalculate the width and height of the grid cells when the window is resized.