Here are the changes I made on css for mobile responsiveness:
1.Add media query of min-width:430px to lessen
the margin size.

2. Adjusted the marigin size on my media query of
min-width: 540px to make it more responsive.

3. Adjusted the width and height of the grid cells
to make it more responsive on the media query of
min-width: 540px.

4. Adjusted the width and height of the buttons
to make it more responsive on the media query of min-width: 540px.

5. Adjusted the padding of the .tools class on the media query
of min-width: 768px to make it more responsive on the tablet size.

6. Set the tools class to position: fixed so that the buttons will remain on
the bottom of the screen when viewed on mobile devices.